Update canvas examples

+++ examples/Canvas/Fancy lines.kt
@@ -1,4 +1,20 @@
 /*
+ * Copyright 2000-2014 JetBrains s.r.o.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
 This example is based on example from html5 canvas2D docs:
   http://www.w3.org/TR/2dcontext/
 Note that only a subset of the api is supported for now.
@@ -6,9 +22,9 @@ Note that only a subset of the api is supported for now.
 
 package fancylines
 
-import js.dom.html5.*
-import js.dom.html.window
-import js.jquery.*
+import kotlin.js.dom.html.window
+import kotlin.js.dom.html5.*
+import jquery.*
 
 fun main(args: Array<String>) {
   jq {
@@ -41,9 +57,9 @@ class FancyLines() {
     y = height * Math.random();
 
     context.bezierCurveTo(width * Math.random(), height * Math.random(),
-        width * Math.random(), height * Math.random(), x, y);
+                width * Math.random(), height * Math.random(), x, y);
 
-    hue += Math.random() * 10;
+    hue += (Math.random() * 10).toInt();
 
     context.strokeStyle = "hsl($hue, 50%, 50%)";
 
@@ -61,7 +77,7 @@ class FancyLines() {
   }
 
   fun run() {
-    window.setInterval({line()}, 40);
-    window.setInterval({blank()}, 100);
+    window.setInterval({ line() }, 40);
+    window.setInterval({ blank() }, 100);
   }
 }
